[TWINFACES-458] feat: update routing for links and create separete hook for creating link
