+++ src/entities/link/api/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from "./use-fetch-by-id";
 export * from "./use-search-v1";
+export * from "./use-create";
